Update Gradle build
This commit upgrades to Gradle 6.8.3 and makes some required changes
in order to prepare for Gradle 7 (in particular, get rid of the old
`compile` dependencies).

It upgrades the publishing mechanism to use the first-class support
for generating sources/javadocs.

In addition, this integrates with the toolchain support, simplifying
the build and making sure that a Java 8 JVM is used for building and
testing.